use gperf for the pluginsd parser
