+++ libnetdata/Makefile.am
@@ -20,7 +20,6 @@ SUBDIRS = \
     locks \
     log \
     onewayalloc \
-    parser \
     popen \
     procfile \
     simple_pattern \
